+++ ipwhois/experimental.py
@@ -158,11 +158,14 @@ def bulk_lookup_rdap(addresses=None, inc_raw=False, retry_count=3, depth=0,
                 'ip_lookup_total' (int) - The total number of addresses that
                     lookups were attempted for, excluding any that failed ASN
                     registry checks.
+                'ip_failed_total' (int) - The total number of addresses that
+                    lookups failed for. Excludes any that failed initially, but
+                    succeeded after further retries.
                 'lacnic' (dict) -
                 {
                     'failed' (list) - The addresses that failed to lookup.
                         Excludes any that failed initially, but succeeded after
-                        futher retries.
+                        further retries.
                     'rate_limited' (list) - The addresses that encountered
                         rate-limiting. Unless an address is also in 'failed',
                         it eventually succeeded.
@@ -196,6 +199,7 @@ def bulk_lookup_rdap(addresses=None, inc_raw=False, retry_count=3, depth=0,
         'ip_input_total': len(addresses),
         'ip_unique_total': 0,
         'ip_lookup_total': 0,
+        'ip_failed_total': 0,
         'lacnic': {'failed': [], 'rate_limited': [], 'total': 0},
         'ripencc': {'failed': [], 'rate_limited': [], 'total': 0},
         'apnic': {'failed': [], 'rate_limited': [], 'total': 0},
@@ -425,6 +429,7 @@ def bulk_lookup_rdap(addresses=None, inc_raw=False, retry_count=3, depth=0,
 
                                 del asn_parsed_results[ip]
                                 stats[rir]['failed'].append(ip)
+                                stats['ip_failed_total'] += 1
 
                                 if rir == 'lacnic':